Support EnableSystemDiagnosticsTracing
It seems the concrete type for ActionDescriptor depends on whether or not the Global Web API Services contains an instance of ITraceWriter (see: Tracing in ASP.NET Web API). By default, the ActionDescriptor will be of type ReflectedHttpActionDescriptor. But when tracing is enabled -- by calling the config.EnableSystemDiagnosticsTracing() -- the ActionDescriptor will be wrapped inside an HttpActionDescriptorTracer type instead.